% Generated by roxygen2 (4.0.1): do not edit by hand
\name{confsand}
\alias{confsand}
\title{Robust confidence levels for GLM object}
\usage{
confsand(x)
}
\arguments{
\item{x}{glm object}
}
\value{
Confidence invervals
}
\description{
Robust confidence levels for GLM object
}
\examples{
require(sandwich)
 counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)

glm.D <- glm(counts ~ outcome + treatment, family=poisson())
confsand(glm.D)
}

