\name{CosmoGLM-package}
\alias{CosmoGLM-package}
\alias{CosmoGLM}
\docType{package}
\title{
Functions and Datasets for GLM analysis in Astronomy and Cosmology
~~ CosmoGLM ~~
}
\description{
This package provides functions and datasets for  Generalized Linear Models applications in Astronomy. Tools  for fitting, cross-validation and displaying the results are also provided.  
}
\details{
\tabular{ll}{
Package: \tab CosmoGLM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-21\cr
License: \tab GPL-3\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Rafael S. de Souza

Maintainer: Rafael S. de Souza <rafael.2706@gmail.com>
}
\keyword{ package }
\seealso{
The \code{ml.nb} function in the \code{COUNT} package
}

