% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggplot.glm}
\alias{ggplot.glm}
\title{ggplot of GLM diagnostics}
\usage{
ggplot.glm(x)
}
\arguments{
\item{x}{A GLM object}
}
\value{
plot
}
\description{
ggplot of GLM diagnostics
}
\examples{
require(ggplot2)
 counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)

glm.D <- glm(counts ~ outcome + treatment, family=poisson())
ggplot.glm(glm.D)
}

