% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_Prob}
\alias{plot_Prob}
\title{ggplot of probabilities for glm object}
\usage{
plot_Prob(glm)
}
\arguments{
\item{glm}{glm  object}
}
\value{
plot
}
\description{
ggplot of probabilities for glm object
}
\examples{
require(ggplot2)
require(scales)
require(ggthemes)
 counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)

glm.D <- glm(counts ~ outcome + treatment, family=poisson())
plot_Prob(glm.D)
}

