% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotProbmcmc}
\alias{plotProbmcmc}
\title{ggplot of probabilities for MCMC object}
\usage{
plotProbmcmc(mcmcglm)
}
\arguments{
\item{mcmcglm}{mcmc  object}
}
\value{
plot
}
\description{
ggplot of probabilities for MCMC object
}
\examples{
require(ggplot2)
require(scales)
require(ggthemes)
require(MCMCpack)
 response <- c(0,1,0,0,1,1,1,0,1)
outcome <- gl(3,1,9)
treatment <- gl(3,3)

glm.D <- MCMClogit(response ~ outcome + treatment, family=poisson())
plotProbmcmc(glm.D)
}

